--- client/custom_validations/validators.js
@@ -1,6 +1,0 @@
-/*eslint-disable*/
-export function isNotBlankSpace(value) {
-  if (value)
-    return value.replace(/\s/g,'')!='';
-  return false;
-}
